// src/api/generated/createProductRequest.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Arannati API
 * API gateway for Arannati microservices
 * OpenAPI spec version: 1.0
 */

export interface CreateProductRequest {
	name: string
	slug: string
	summary: string
	description: string
	images: string[]
	price: number
	discountAmount?: number
	currency: string
	countryOfOrigin: string
	brandId: string
	categoryIds: string[]
}